% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area_rect}
\alias{area_rect}
\title{Area Under the Curve}
\usage{
area_rect(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\value{
A numeric.
}
\description{
Compute area under the curve by using the rectangle's method.
}
\examples{
x <- c(0, 0, 0.5, 1)
y <- c(0, 0.5, 0.5, 1)
plot(x, y, type = "b")
evabic:::area_rect(x, y)
}
