% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\docType{data}
\name{ebc_allmeasures}
\alias{ebc_allmeasures}
\title{Supported measures}
\format{An object of class \code{character} of length 15.}
\usage{
ebc_allmeasures
}
\description{
Supported measures
}
\keyword{datasets}
