% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_measures.R
\name{ebc_confusion}
\alias{ebc_confusion}
\title{Confusion matrix}
\usage{
ebc_confusion(detected, true, all, m = length(all), prop = FALSE)
}
\arguments{
\item{detected}{Vector of elements that are detected.}

\item{true}{Vector of element that are supposed to be detected.}

\item{all}{Vector of all elements.}

\item{m}{Total number of elements.}

\item{prop}{Logical, default to \code{FALSE}. Should the matrix sum to one?}
}
\value{
A 2*2 named matrix.
}
\description{
Compute the the confusion matrix
}
\details{
See \code{\link{ebc_allmeasures}} for the description of the measures.
}
\examples{
ebc_confusion(detected = c("A", "C", "D"), true = c("A", "B", "C"), m = 6)
}
