% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_measures.R
\name{ebc_TPR}
\alias{ebc_TPR}
\alias{ebc_TNR}
\alias{ebc_FDR}
\alias{ebc_ACC}
\title{Derived measures.}
\usage{
ebc_TPR(detected, true)

ebc_TNR(detected, true, all, m = length(all))

ebc_FDR(detected, true)

ebc_ACC(detected, true, all, m = length(all))
}
\arguments{
\item{detected}{Vector of elements that are detected.}

\item{true}{Vector of element that are supposed to be detected (\eqn{H1}).}

\item{all}{Vector of all elements.}

\item{m}{Total number of element.}
}
\value{
A numeric.
}
\description{
Measures derived from confusion matrix.
}
