% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nl2c}
\alias{nl2c}
\title{Named logical to character}
\usage{
nl2c(x)
}
\arguments{
\item{x}{vector}
}
\value{
a vector of the same size
}
\description{
Extracts names of \code{x} where \code{x} is \code{TRUE}
}
\examples{
x <- c(a = TRUE, b = FALSE, c = FALSE, d = TRUE)
evabic:::nl2c(x)

}
\keyword{internal}
