% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_measures.R
\name{ebc_TPR}
\alias{ebc_TPR}
\alias{ebc_TNR}
\alias{ebc_PPV}
\alias{ebc_FDR}
\alias{ebc_ACC}
\alias{ebc_F1}
\title{Derived measures.}
\usage{
ebc_TPR(detected, true)

ebc_TNR(detected, true, all, m = length(all))

ebc_PPV(detected, true)

ebc_FDR(detected, true)

ebc_ACC(detected, true, all, m = length(all))

ebc_F1(detected, true)
}
\arguments{
\item{detected}{Vector of elements that are detected.}

\item{true}{Vector of element that are supposed to be detected (\eqn{H1}).}

\item{all}{Vector of all elements.}

\item{m}{Total number of element.}
}
\value{
A numeric.
}
\description{
Measures derived from confusion matrix.
}
\details{
TPR: True Positive Rate, also known as Sensitivity and Recall.

TNR: True Negative Rate, also known as Specificity and Selectivity.

PPV: Positive Predictive Value, also known as Precision.

FDR: False Discovery Rate.

ACC: Accuracy.

F1: F1 score.
}
\examples{
ebc_TPR(detected = c("A", "C", "D"), true = c("A", "B", "C"))
ebc_ACC(detected = c("A", "C", "D"), true = c("A", "B", "C"),
        all = LETTERS[1:6])
}
